adding bits
