+++ tests/test_func_equivalence.py
@@ -4,3 +4,7 @@ Tests that rust functions and python function return the same values
 
 from helper import set_wd
 set_wd()
+
+from core
+
+function_pairs = [()]
